[Improve] Streampark-common module tika dependency moved to the streampark-console module
