+++ streampark-common/pom.xml
@@ -156,13 +156,6 @@
             <optional>true</optional>
         </dependency>
 
-        <!--tika -->
-        <dependency>
-            <groupId>org.apache.tika</groupId>
-            <artifactId>tika-core</artifactId>
-            <version>1.20</version>
-        </dependency>
-
     </dependencies>
 
     <build>
