+++ streampark-common/src/main/scala/org/apache/streampark/common/util/FileUtils.scala
@@ -18,12 +18,6 @@ package org.apache.streampark.common.util
 
 import org.apache.streampark.common.util.ImplicitsUtils._
 
-import org.apache.commons.lang3.StringUtils
-import org.apache.tika.metadata.{HttpHeaders, Metadata}
-import org.apache.tika.mime.MediaType
-import org.apache.tika.parser.{AutoDetectParser, ParseContext}
-import org.xml.sax.helpers.DefaultHandler
-
 import java.io._
 import java.net.URL
 import java.nio.ByteBuffer
@@ -62,20 +56,6 @@ object FileUtils {
         in.read(b, 0, b.length)
         bytesToHexString(b)
       }) == "504B0304"
-  }
-
-  def isPythonFileType(contentType: String, input: InputStream): Boolean = {
-    if (StringUtils.isBlank(contentType) || input == null) {
-      throw new RuntimeException("The contentType or inputStream can not be null")
-    }
-    getMimeType(input) == MediaType.TEXT_PLAIN.toString && contentType.contains("text/x-python")
-  }
-
-  def getMimeType(stream: InputStream): String = {
-    val metadata: Metadata = new Metadata
-    val parser: AutoDetectParser = new AutoDetectParser
-    parser.parse(stream, new DefaultHandler, metadata, new ParseContext)
-    metadata.get(HttpHeaders.CONTENT_TYPE)
   }
 
   def isJarFileType(file: File): Boolean = {
